Downsample reads in a 10X Genomics dataset

Generate a UMI count matrix after downsampling reads from the molecule information file produced by CellRanger for 10X Genomics data.

downsampleReads(
  sample,
  prop,
  barcode.length = NULL,
  bycol = FALSE,
  features = NULL,
  use.library = NULL
)

Arguments

sample
A string containing the path to the molecule information HDF5 file.

prop
A numeric scalar or, if bycol=TRUE, a vector of length ncol(x). All values should lie in [0, 1] specifying the downsampling proportion for the matrix or for each cell.

barcode.length
An integer scalar specifying the length of the cell barcode, see read10xMolInfo.

bycol
A logical scalar indicating whether downsampling should be performed on a column-by-column basis.

features
A character vector containing the names of the features on which to perform downsampling.

use.library
An integer vector specifying the library indices for which to extract molecules from sample. Alternatively, a character vector specifying the library type(s), e.g., "Gene expression".

Value

A numeric sparse matrix containing the downsampled UMI counts for each gene (row) and barcode (column). If features is set, only the rows with names in features are returned.

Details

This function downsamples the reads for each molecule by the specified prop, using the information in sample. It then constructs a UMI count matrix based on the molecules with non-zero read counts. The aim is to eliminate differences in technical noise that can drive clustering by batch, as described in downsampleMatrix.

Subsampling the reads with downsampleReads recapitulates the effect of differences in sequencing depth per cell. This provides an alternative to downsampling with the CellRanger aggr function or subsampling with the 10X Genomics R kit. Note that this differs from directly subsampling the UMI count matrix with downsampleMatrix.

If bycol=FALSE, downsampling without replacement is performed on all reads from the entire dataset. The total number of reads for each cell after downsampling may not be exactly equal to prop times the original value. Note that this is the more natural approach and is the default, which differs from the default used in downsampleMatrix.

If bycol=TRUE, sampling without replacement is performed on the reads for each cell. The total number of reads for each cell after downsampling is guaranteed to be prop times the original total (rounded to the nearest integer). Different proportions can be specified for different cells by setting prop to a vector, where each proportion corresponds to a cell/GEM combination in the order returned by get10xMolInfoStats.

The use.library argument is intended for studies with multiple feature types, e.g., antibody capture or CRISPR tags. As the reads for each feature type are generated in a separate sequencing library, it is generally most appropriate to downsample reads for each feature type separately. This can be achieved by setting use.library to the name or index of the desired feature set. The features of interest can also be directly specified with features. (This will be intersected with any use.library choice if both are specified.)

See also

Author

Aaron Lun

Examples

# Mocking up some 10X HDF5-formatted data.
out <- DropletUtils:::simBasicMolInfo(tempfile())

# Downsampling by the reads.
downsampleReads(out, barcode.length=4, prop=0.5)

#> 20 x 256 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 256 column names 'AAAA-1', 'AAAC-1', 'AAAG-1' ... ]]

#>                                                                               
#> ENSG1  4 2 2 1 . 1 5 1 1 3 . 3 . 3 1 7 2 2 2 2 3 6 2 1 3 . 2 2 1 2 4 3 4 1 4 2
#> ENSG2  . 1 . 1 2 . 3 . 1 4 1 . 2 1 1 1 . 6 4 1 3 . 1 1 3 3 3 4 3 1 2 3 1 2 2 .
#> ENSG3  3 1 3 3 1 4 3 2 1 2 1 5 1 4 5 3 5 1 2 2 3 2 1 . 3 3 . 5 1 . 2 2 1 2 2 .
#> ENSG4  3 1 2 2 1 . 1 2 2 2 3 2 2 2 2 3 2 2 . 1 3 3 1 1 3 5 1 3 . 4 2 . 2 3 1 2
#> ENSG5  2 3 1 1 5 1 1 2 2 1 . 2 . 1 . 2 2 . 2 1 1 . 2 2 1 1 1 2 2 3 4 1 4 . . 3
#> ENSG6  1 . 2 3 3 2 1 3 1 2 1 . . 5 1 . 4 1 2 3 2 . 4 1 1 2 2 2 4 . 2 5 1 . 1 1
#> ENSG7  . 2 2 1 1 1 . 1 3 . 3 1 1 5 1 2 1 1 2 6 3 1 . 1 2 . 1 3 2 2 2 3 1 4 1 1
#> ENSG8  . . 3 3 2 . . 2 2 4 2 3 1 2 4 1 3 2 3 2 . 4 4 2 1 2 1 2 . 6 3 . 1 3 1 1
#> ENSG9  1 . 1 2 6 1 1 2 1 3 3 2 4 2 1 3 1 2 1 5 1 3 1 . . 3 3 2 3 3 . . 4 4 1 .
#> ENSG10 . 5 1 1 3 1 2 1 . . 4 1 2 . . 1 1 7 1 1 . 1 . 4 . 4 1 . . 2 1 1 2 2 1 1
#> ENSG11 . 2 2 3 4 2 2 2 2 1 1 4 1 3 3 1 1 1 3 . 2 . 1 2 2 2 2 . . 1 3 2 1 1 1 1
#> ENSG12 2 1 1 . 4 2 3 2 3 3 4 . 2 1 4 1 1 2 4 3 4 1 . . 6 2 4 1 4 3 1 1 1 6 2 5
#> ENSG13 1 2 5 2 4 2 2 1 1 . 2 . 1 2 1 1 1 2 1 3 1 1 2 3 4 2 3 4 3 2 2 5 2 3 2 3
#> ENSG14 1 . 1 2 2 . 3 1 1 . 3 1 2 1 . 1 4 4 1 2 2 2 2 3 3 . . 1 2 . 1 1 2 1 1 .
#> ENSG15 3 . 2 . . 2 2 1 3 . 3 1 2 1 1 . 1 6 2 2 1 1 . 1 6 2 . 1 4 1 3 1 2 2 . .
#> ENSG16 3 1 1 2 1 3 1 . 2 2 2 . . 4 1 1 2 1 1 2 3 4 . 3 2 2 4 2 1 2 1 2 1 2 4 .
#> ENSG17 2 3 4 1 5 1 2 3 4 3 1 1 2 . . 2 3 1 1 3 . 1 4 1 5 . 7 2 4 1 1 3 3 1 4 3
#> ENSG18 1 4 3 4 4 1 2 1 2 3 2 4 3 3 . 2 7 . 2 1 1 1 . 5 2 2 3 1 3 2 1 1 1 3 3 .
#> ENSG19 . 2 3 1 1 1 . 4 2 . 4 . 1 1 1 3 3 1 2 3 2 3 2 2 1 2 . . 2 2 1 2 3 2 3 1
#> ENSG20 2 3 . 1 . 4 2 4 3 1 . 2 4 1 . 2 3 1 1 1 2 1 3 . 2 3 1 2 . 6 3 2 3 . 2 1
#>                                                                               
#> ENSG1  4 . 1 1 . . 4 2 2 4 3 3 2 1 2 3 2 3 5 1 1 1 . . 3 3 2 2 3 2 1 1 . 1 1 2
#> ENSG2  4 . 1 1 1 . 1 2 1 . 2 4 2 1 4 . 2 2 . 1 . 1 . 2 1 3 . 2 2 3 2 2 1 2 4 4
#> ENSG3  4 . . 3 1 3 2 1 1 1 . 4 3 1 3 1 . 5 1 1 3 2 2 1 1 4 . 2 2 . 1 1 5 4 4 .
#> ENSG4  2 2 2 2 1 2 3 . 1 1 5 2 3 1 1 1 1 . . 3 3 1 3 1 3 2 5 4 2 3 5 1 2 1 . 3
#> ENSG5  2 1 1 2 . 1 2 2 1 2 . 2 3 2 1 2 2 1 3 2 2 2 1 . 1 . 1 1 2 1 2 2 3 2 1 3
#> ENSG6  . . 1 . 3 1 1 1 3 2 1 . 2 . 2 . 6 1 3 4 . 2 . 1 4 2 1 2 3 1 . 3 3 1 3 1
#> ENSG7  2 1 1 3 4 3 1 1 . 4 1 3 1 1 5 . 3 2 4 3 . 3 . 1 2 1 1 . 2 2 . 1 2 1 2 2
#> ENSG8  4 1 2 1 1 1 4 2 2 4 . 1 1 2 . 5 1 3 2 2 1 1 2 2 2 . 1 1 1 2 2 4 2 2 . 1
#> ENSG9  . 1 2 2 3 3 3 2 . 2 5 6 3 . 4 . 2 1 1 . 1 2 . 2 2 2 1 1 1 . 3 . 1 2 2 3
#> ENSG10 . 3 . 2 1 2 2 1 1 2 4 3 2 3 . 2 1 . 1 . 3 1 3 3 . . 1 3 1 4 1 2 . 3 3 .
#> ENSG11 4 1 1 1 2 2 2 1 2 3 1 . . 1 3 2 1 2 2 4 1 2 . 1 2 2 1 3 2 1 1 . 2 3 2 1
#> ENSG12 2 5 1 4 2 2 1 3 3 3 5 3 2 1 1 2 2 3 1 . 1 4 1 2 3 2 . 1 . . 5 . 2 5 1 2
#> ENSG13 . . 1 1 . 1 2 3 4 2 3 1 1 1 1 1 . 4 3 1 1 1 3 4 1 1 2 . 2 1 4 2 . 4 3 3
#> ENSG14 3 2 2 2 2 1 4 2 2 1 2 3 4 5 1 1 . 3 1 3 2 . 1 . 3 3 2 4 2 3 1 1 1 1 1 4
#> ENSG15 2 1 . 1 2 3 3 6 1 . . 2 . 4 4 2 2 5 1 3 1 1 1 . 3 1 1 5 2 . 3 1 3 2 3 5
#> ENSG16 1 1 3 . 3 2 3 . 2 2 4 6 . 1 . . 2 . 4 1 . 1 4 . 2 1 4 1 4 1 1 5 4 4 3 2
#> ENSG17 3 2 1 2 2 . 2 3 3 . 1 1 4 3 1 . 1 3 3 1 3 2 2 4 . 1 1 1 3 3 5 1 1 3 3 3
#> ENSG18 1 3 3 2 3 . 1 2 2 . 1 1 . 2 2 1 1 2 1 . 3 2 4 3 4 3 . 3 2 3 4 4 1 . 4 .
#> ENSG19 2 2 2 4 3 1 1 4 3 1 2 4 5 3 1 3 . 1 4 3 3 3 1 4 2 2 4 1 3 1 1 1 1 1 1 .
#> ENSG20 3 . . 1 2 1 2 . 7 4 1 1 2 2 2 2 1 2 3 2 . 3 5 2 3 1 1 . 3 1 2 . 4 3 1 3
#>                                                                               
#> ENSG1  1 . . 4 . 5 2 2 2 2 1 3 3 2 3 1 2 . 1 . . . 2 1 2 2 . 1 3 4 . 2 2 4 2 2
#> ENSG2  3 1 3 2 3 1 . 3 1 1 1 6 1 5 1 2 2 3 3 1 . . 1 1 1 4 2 1 3 1 2 3 2 3 2 3
#> ENSG3  2 2 1 3 1 4 1 5 3 2 6 . . . 3 2 4 1 3 2 . . . 2 . 2 1 2 2 3 . 3 . 1 1 4
#> ENSG4  . 1 1 1 . . 4 . 2 1 2 1 1 3 1 3 2 1 4 4 6 . 3 3 3 2 2 1 1 2 . . 1 1 2 5
#> ENSG5  4 1 2 1 2 . 5 1 4 . 6 1 2 2 2 1 2 3 2 2 1 . . 3 3 4 3 3 2 4 1 2 1 1 4 1
#> ENSG6  1 2 1 3 1 1 3 1 . 1 1 1 2 3 2 1 2 . 1 3 1 2 . 1 1 3 2 3 2 3 1 . 1 3 3 1
#> ENSG7  3 3 1 3 4 2 1 1 . 2 1 . . 1 2 2 2 1 3 3 2 . 2 4 5 1 4 2 2 1 2 2 2 2 2 2
#> ENSG8  . . 2 2 3 2 3 1 . 2 3 3 . 1 3 1 2 . 1 3 1 3 4 1 3 2 1 2 . 1 2 2 2 3 2 5
#> ENSG9  . . . 5 . 3 1 1 . 2 1 1 2 3 1 3 . 5 1 2 2 . 3 3 . 2 1 2 1 5 2 3 3 1 3 .
#> ENSG10 3 2 . 1 2 3 2 2 1 2 3 4 1 3 1 3 2 2 1 1 1 3 1 3 1 . 1 3 4 2 1 4 2 2 1 .
#> ENSG11 2 2 3 2 . 2 3 1 3 1 3 4 3 2 1 . . 4 2 3 2 . 1 4 2 1 2 2 1 2 2 3 . 2 1 1
#> ENSG12 1 2 4 . . 2 5 4 2 2 3 1 1 2 3 4 2 3 2 1 . . 3 2 3 1 . 2 . . . . 7 1 3 2
#> ENSG13 . 2 3 2 3 . 2 1 4 . 1 . 2 2 . 1 . 1 3 2 2 3 3 . 2 8 1 2 3 1 . 1 2 1 1 2
#> ENSG14 2 1 3 1 1 3 1 2 1 3 3 2 2 1 . 2 2 . 3 3 2 1 3 . 2 3 2 1 3 3 . 3 2 2 1 2
#> ENSG15 1 4 1 4 1 2 2 2 3 2 2 . 2 2 2 . 2 1 2 2 1 5 4 1 3 2 3 2 1 3 5 1 1 1 3 .
#> ENSG16 1 2 1 1 2 1 2 . 2 2 . 4 1 1 . 2 1 3 2 3 3 4 1 2 2 2 . 2 . 2 1 3 2 1 2 .
#> ENSG17 1 2 . . 3 3 1 . . 2 3 3 . 2 4 1 1 1 4 1 6 . 1 1 . 1 2 5 . 6 . 3 1 1 2 1
#> ENSG18 3 4 1 . 2 2 . 3 1 1 2 3 3 . 2 1 2 2 1 2 5 3 4 3 . 2 2 4 3 1 6 1 2 . 2 3
#> ENSG19 1 . 3 1 4 3 1 3 2 3 2 4 . 2 . 1 3 . 2 2 1 8 . 2 3 1 1 2 2 2 1 . 1 1 2 1
#> ENSG20 3 2 3 2 2 2 2 2 2 2 3 . 3 2 2 1 2 . 3 . 3 1 . 1 2 4 2 1 2 3 4 4 1 4 3 1
#>                                                                               
#> ENSG1  1 2 . . . 2 1 1 2 1 1 1 2 2 1 . 6 2 1 . 3 . 2 1 4 . 4 1 1 2 3 3 2 3 3 .
#> ENSG2  . 3 . 1 . 1 2 3 3 3 2 1 . . 4 5 1 2 6 . 3 . 1 2 2 1 1 3 2 1 1 3 3 1 1 2
#> ENSG3  4 3 1 . 4 1 1 2 3 4 3 1 1 1 1 4 5 1 1 4 . 2 5 3 4 1 1 2 2 3 . . 3 1 . 1
#> ENSG4  2 . . 2 1 . 2 2 1 1 . 5 2 2 1 2 1 3 2 2 1 5 1 1 3 1 4 5 . 3 . 2 . 1 1 3
#> ENSG5  1 2 . 1 2 4 2 2 4 2 1 2 5 3 3 3 . . 2 5 4 3 1 1 2 2 2 2 3 1 4 . . 1 1 5
#> ENSG6  2 2 1 1 2 . 1 5 2 3 3 3 . 2 . 1 . 6 1 3 1 4 3 3 1 3 3 2 3 3 . 5 1 2 2 .
#> ENSG7  1 . 1 1 2 4 1 . 4 1 3 3 3 . 2 1 3 7 3 1 2 1 2 3 1 6 2 3 3 . . 1 . 3 . 1
#> ENSG8  2 1 3 1 2 4 1 3 1 1 2 2 . 1 3 1 2 3 . 1 1 4 2 . 1 3 3 4 1 2 1 2 3 1 3 3
#> ENSG9  1 1 2 . 3 3 1 5 . 3 1 1 . 2 3 2 1 . 3 1 1 8 1 1 3 1 . . 2 1 . 1 1 1 1 5
#> ENSG10 2 3 . 1 3 2 . 2 1 . 2 2 2 1 4 1 2 . 3 1 1 . 4 . 1 2 1 . 2 3 2 2 2 1 1 1
#> ENSG11 3 . 3 2 3 2 2 3 2 1 1 1 2 . . 1 2 3 . 3 1 1 3 5 4 2 2 4 4 2 . 2 1 4 3 2
#> ENSG12 2 2 3 . . 1 1 . 1 4 3 . 4 4 4 1 2 1 2 . 3 . . 2 1 1 3 5 3 . 2 2 2 1 2 2
#> ENSG13 2 2 3 4 2 4 6 3 3 3 . . 3 3 3 3 2 . 2 1 . 3 2 1 . 3 2 . . 4 . 3 2 1 2 4
#> ENSG14 . 3 1 2 4 5 2 2 . 4 3 3 3 . 2 3 3 3 1 3 2 1 4 . 4 . . 1 1 4 . 3 . 1 1 2
#> ENSG15 1 1 4 2 . 1 1 1 1 1 5 3 1 5 2 . 3 2 . 2 1 1 3 3 3 4 1 2 1 3 3 4 1 2 . 1
#> ENSG16 . 2 3 3 3 3 1 . 2 1 4 2 2 2 1 2 2 . 2 2 . 3 . 2 2 1 . . 1 2 3 . 2 1 2 3
#> ENSG17 1 . 1 2 . 4 3 . 1 2 2 1 2 . 3 3 2 1 1 2 3 2 . 3 2 3 2 1 1 1 . 1 2 . 1 .
#> ENSG18 2 1 3 2 3 4 4 3 2 . 2 . 2 4 2 . . 3 3 . . 1 . 3 1 3 3 1 4 1 2 2 . 1 1 2
#> ENSG19 2 2 1 . 4 . 5 1 2 1 1 4 . 2 . 5 1 2 . 1 4 2 5 1 3 4 2 . 2 2 2 1 . 2 3 1
#> ENSG20 2 3 1 2 . 1 . 2 6 1 2 4 . 4 2 1 4 1 2 2 3 2 4 1 6 . . 3 . 4 3 3 1 . 2 2
#>                                                                               
#> ENSG1  1 1 3 1 1 2 . 1 3 . 2 2 1 2 2 1 1 2 4 1 2 1 1 4 . 3 2 2 4 2 1 4 3 . 2 5
#> ENSG2  1 2 2 3 1 2 2 1 1 3 . 2 2 . 1 2 . 1 3 2 2 2 3 1 3 . 1 3 4 1 2 1 1 . 5 1
#> ENSG3  4 1 4 2 2 2 4 1 5 . 4 . 2 . 1 . 3 3 2 2 . 1 2 1 1 . 1 2 . 4 1 1 2 3 1 2
#> ENSG4  1 2 3 1 3 1 2 2 4 3 2 . 3 . 1 1 1 . 1 . 3 2 2 1 2 . 2 1 3 4 . . 1 1 1 1
#> ENSG5  1 2 1 3 1 1 1 1 2 3 . 3 2 6 3 1 3 2 . 5 2 2 4 3 2 1 1 3 . 3 2 1 2 4 . 2
#> ENSG6  . 1 2 2 1 . 2 . 1 1 1 1 . 2 3 3 1 1 1 1 1 3 4 . 1 1 1 3 2 2 1 . . 1 1 .
#> ENSG7  2 5 1 1 3 4 2 2 5 2 . 3 1 2 2 3 3 4 2 2 3 2 1 1 5 1 1 . 1 1 2 3 . 3 1 2
#> ENSG8  2 . 4 . 2 2 1 2 1 . . 1 3 1 1 3 1 2 2 1 2 1 . 2 1 . 2 . 2 1 1 1 3 1 2 3
#> ENSG9  1 1 3 2 3 . 2 1 1 2 4 1 2 6 . 1 3 2 2 2 2 2 . 1 4 1 3 2 2 3 . 2 5 . 3 3
#> ENSG10 3 1 . . 3 . 2 2 1 1 1 1 1 4 1 3 2 . 1 3 2 2 1 4 2 2 2 4 3 2 3 1 1 1 3 1
#> ENSG11 1 3 . 1 . 2 1 6 . 1 3 2 2 1 1 . 1 3 5 2 1 . 3 3 . . 1 2 2 3 1 2 . 1 1 4
#> ENSG12 . 1 . 3 1 . 3 4 . . 2 1 4 1 1 . 1 . 2 4 1 6 1 1 2 2 3 2 2 4 1 . . 2 3 1
#> ENSG13 . . 5 2 5 4 1 1 3 3 1 1 1 1 . 2 1 2 . 1 2 2 1 4 1 3 . . 3 . 2 . 1 3 3 .
#> ENSG14 3 5 1 1 1 3 1 . 1 6 1 . 4 2 1 . 1 2 1 3 1 1 3 2 1 1 1 4 4 2 2 . 6 1 . 2
#> ENSG15 . 1 1 3 1 1 3 2 . . 2 1 4 2 1 3 . 1 1 5 3 1 3 . 1 . 1 2 . 3 4 . 4 . . 1
#> ENSG16 . . 2 2 3 . 2 . 2 . 2 . . 2 2 1 1 1 . 1 4 1 1 3 1 2 . 3 1 1 2 3 3 . 2 1
#> ENSG17 1 3 1 2 3 2 1 4 2 1 3 2 2 3 3 4 2 1 3 1 5 2 3 1 5 2 1 2 . 1 1 2 . . 1 2
#> ENSG18 4 2 3 1 . 1 . 1 1 2 2 1 1 . 5 1 1 4 1 2 2 2 4 1 . 1 2 3 1 1 1 3 2 2 4 3
#> ENSG19 5 1 2 . 2 2 2 . 3 . 2 3 4 . 1 1 1 2 1 2 1 3 1 1 1 2 1 . 2 1 1 2 3 3 2 4
#> ENSG20 . 1 1 . 2 . 3 2 4 . 1 3 2 1 5 . 3 3 1 1 2 5 1 4 2 1 3 2 1 5 1 . . 2 . 2
#>                                                                               
#> ENSG1  1 3 . 1 4 2 3 1 1 1 1 4 2 1 . 2 . 3 1 1 . 2 3 1 2 1 1 3 3 2 . 1 3 4 4 1
#> ENSG2  3 3 4 2 1 1 . 3 2 4 . 3 3 2 1 4 . 4 2 2 3 2 1 1 2 1 2 2 . 2 2 1 2 6 1 3
#> ENSG3  . 2 3 1 4 2 3 4 5 1 . 1 1 2 5 1 1 1 3 4 1 1 1 2 5 1 4 5 2 1 1 1 2 3 1 3
#> ENSG4  3 . 2 4 1 3 3 2 1 1 7 4 . 5 . . 2 1 3 1 6 2 3 2 1 1 2 2 1 2 2 2 . 3 . 2
#> ENSG5  1 3 2 1 3 1 1 1 1 5 2 2 . 1 . 2 3 3 2 1 3 2 1 1 4 1 3 2 . 3 5 . 2 4 3 3
#> ENSG6  2 3 1 2 4 1 2 2 1 2 3 3 . . 2 . . 1 1 2 1 4 . 1 1 1 2 2 2 2 2 . 1 2 1 1
#> ENSG7  3 1 3 3 4 1 1 3 2 5 . 3 2 1 2 2 1 1 2 2 1 2 2 2 4 1 3 2 2 . 5 . 3 4 2 6
#> ENSG8  1 1 1 2 5 1 3 . 1 1 4 2 3 2 2 . . 2 3 2 2 1 3 3 3 1 3 1 1 4 . 2 2 4 1 2
#> ENSG9  4 1 2 3 2 2 1 1 1 . 5 1 1 1 1 1 1 2 3 1 . 2 1 1 1 . 1 1 1 1 2 1 1 . 2 2
#> ENSG10 1 1 1 . . 2 1 . 3 1 2 4 . 3 4 1 . 2 1 2 2 2 2 1 1 2 1 2 . 2 2 2 3 3 1 1
#> ENSG11 1 2 . 2 1 . 2 1 2 . 3 . 1 3 2 . 1 1 3 1 1 . 2 1 2 2 1 2 . 1 2 . 4 3 4 2
#> ENSG12 2 3 3 4 1 1 . 1 1 4 1 3 2 . 1 . 7 4 . 2 2 2 . 1 1 4 4 6 1 3 4 2 3 1 3 .
#> ENSG13 1 . 1 2 2 1 2 2 2 1 2 1 1 5 2 . 3 2 . 2 1 1 . . . 1 2 . 1 3 3 1 1 1 . 2
#> ENSG14 3 3 4 3 5 1 . 2 . 2 . 2 . 1 1 2 1 2 . 1 1 3 2 3 4 2 3 . . 1 3 2 . 2 4 3
#> ENSG15 2 1 1 3 3 4 2 1 1 1 3 2 1 1 1 6 1 2 1 2 1 3 . 3 3 1 . 1 7 3 2 5 3 3 1 1
#> ENSG16 2 3 . 2 2 1 1 3 3 1 3 2 2 4 1 1 1 2 . 2 1 1 2 1 3 . 1 2 2 5 4 2 2 2 1 4
#> ENSG17 4 1 3 . 1 1 2 . 5 3 . 3 1 3 3 2 2 3 3 3 4 2 . 1 . . . 1 1 3 3 3 4 1 1 1
#> ENSG18 2 3 2 2 3 3 2 3 2 3 2 3 5 2 3 1 . 2 1 . 2 1 3 6 . 2 2 3 2 3 2 2 4 3 1 2
#> ENSG19 . 6 1 2 2 2 . 3 2 1 1 2 . 2 2 2 4 1 1 2 . 1 3 3 2 3 3 4 . 4 1 1 2 2 3 3
#> ENSG20 2 . 2 4 2 2 1 1 2 3 . 2 4 4 . . 2 2 1 4 2 2 2 3 3 2 4 3 1 3 . 1 1 3 2 2
#>                                                                               
#> ENSG1  3 . 2 1 . . 4 2 2 . 6 2 1 1 1 . 1 2 1 2 1 1 3 2 2 2 2 3 . 2 . 4 . . 2 .
#> ENSG2  2 3 1 . 2 2 1 3 2 1 2 1 1 1 2 2 2 9 3 3 1 2 3 2 . . 3 1 2 1 1 1 4 4 1 4
#> ENSG3  3 2 6 2 . 1 2 2 3 2 3 1 2 . 1 2 2 1 2 1 1 1 1 2 2 3 1 2 . 1 1 2 3 1 . 2
#> ENSG4  1 1 2 3 2 2 2 2 4 4 2 5 4 1 3 2 . 2 . 2 1 3 4 4 4 1 3 2 1 3 2 3 2 3 3 2
#> ENSG5  4 2 1 3 2 2 1 2 1 3 3 1 . 1 2 . 3 4 3 1 1 1 6 . 2 . 1 2 2 1 2 3 1 1 3 3
#> ENSG6  . 3 . 1 1 2 2 2 . 2 2 1 1 2 2 . 2 1 1 2 1 1 . 5 1 1 . 4 1 3 2 1 3 . 1 1
#> ENSG7  4 4 4 2 1 . 1 4 . . 2 . 3 1 2 1 3 . 2 . 2 1 5 1 3 1 3 1 3 1 2 3 1 . . 3
#> ENSG8  4 1 . 2 1 . . 3 1 5 1 . 2 . 2 2 3 3 2 1 3 1 2 1 6 1 2 . 2 1 1 1 1 1 2 .
#> ENSG9  3 2 1 2 1 2 1 1 4 2 3 1 1 3 3 2 1 3 2 7 . 1 2 2 1 2 4 4 1 4 2 3 . 4 3 1
#> ENSG10 . 4 2 1 1 1 1 1 3 3 1 . 4 3 2 . 2 3 1 3 . 4 1 1 3 3 2 1 . 2 3 1 4 . 1 2
#> ENSG11 2 6 2 1 . 3 1 3 2 2 3 1 3 4 2 3 2 1 . 2 1 3 1 . 3 5 2 1 3 1 3 3 3 . 1 3
#> ENSG12 1 1 1 . 4 3 2 1 4 1 . 4 2 2 3 1 1 2 1 4 1 3 . 4 1 2 2 4 1 4 5 5 2 1 . 2
#> ENSG13 2 1 2 2 3 2 . 3 3 3 2 2 . 2 3 . 1 2 3 1 2 1 1 4 2 3 1 6 3 4 1 1 3 4 . 2
#> ENSG14 1 4 1 2 3 1 . . . 1 1 1 4 1 4 . 1 6 2 2 1 1 2 2 1 2 . 2 1 7 2 2 1 1 . 1
#> ENSG15 2 1 3 1 2 . . 3 2 1 2 3 1 3 1 2 1 2 1 2 3 3 . 2 1 2 3 . 2 3 4 2 2 3 4 3
#> ENSG16 . 1 1 . 2 3 1 1 1 1 4 3 . 1 3 5 2 . 3 2 4 . 2 1 3 4 2 1 2 1 3 3 1 1 2 2
#> ENSG17 . 1 4 1 . 3 . 2 1 7 3 . 1 1 3 2 1 . 1 1 7 3 . 1 3 2 5 2 3 2 . . 1 1 1 3
#> ENSG18 1 4 1 1 . 2 2 3 2 . 3 4 2 2 1 4 1 . 5 5 1 5 3 . 1 2 2 1 . 1 2 2 2 . 2 .
#> ENSG19 5 1 3 4 6 . 2 2 . 1 2 1 1 4 1 1 2 2 3 1 2 1 . 2 . 1 3 1 . 2 1 2 1 2 2 1
#> ENSG20 1 1 2 3 2 4 3 5 1 4 3 3 3 1 . 4 2 3 1 . 2 1 . 2 1 2 . 3 5 . . 2 2 2 2 1
#>               
#> ENSG1  1 . 3 .
#> ENSG2  2 3 2 1
#> ENSG3  1 4 4 3
#> ENSG4  2 1 . 1
#> ENSG5  4 3 3 1
#> ENSG6  1 4 4 4
#> ENSG7  . . 4 2
#> ENSG8  3 5 4 2
#> ENSG9  1 3 3 2
#> ENSG10 1 . 1 4
#> ENSG11 2 3 2 2
#> ENSG12 2 1 1 4
#> ENSG13 2 3 1 1
#> ENSG14 . 2 2 1
#> ENSG15 2 2 2 1
#> ENSG16 3 . 2 1
#> ENSG17 2 3 3 2
#> ENSG18 1 2 2 3
#> ENSG19 6 2 2 .
#> ENSG20 2 1 2 3