CellRanger's emptyDrops variant

An approximate implementation of the --soloCellFilter EmptyDrops_CR filtering approach, which itself was reverse-engineered from the behavior of CellRanger 3.

emptyDropsCellRanger(m, ...)

# S4 method for ANY
emptyDropsCellRanger(
  m,
  n.expected.cells = 3000,
  max.percentile = 0.99,
  max.min.ratio = 10,
  umi.min = 500,
  umi.min.frac.median = 0.01,
  can.max.n = 20000,
  ind.min = 45000,
  ind.max = 90000,
  round = TRUE,
  niters = 10000,
  BPPARAM = SerialParam()
)

# S4 method for SummarizedExperiment
emptyDropsCellRanger(m, ..., assay.type = "counts")

Arguments

m
A numeric matrix-like object containing counts, where columns represent barcoded droplets and rows represent features. The matrix should only contain barcodes for an individual sample, prior to any filtering for barcodes.

n.expected.cells
An integer scalar specifying the number of expected cells in a sample.

max.percentile
A numeric scalar between 0 and 1 used to define the maximum UMI count used in the simple filtering algorithm.

max.min.ratio
An integer scalar specifying the ratio of the maximum and minimum UMI count used in the simple filtering algorithm.

umi.min
An integer scalar specifying the minimum UMI count for inclusion of a barcode in the cell candidate pool.

umi.min.frac.median
A numeric scalar between 0 and 1 used to define the minimum UMI count for inclusion of a barcode in the cell candidate pool.

ind.min
An integer scalar specifying the lowest UMI count ranking for inclusion of a barcode in the ambient profile.

ind.max
An integer scalar specifying the highest UMI count ranking for inclusion of a barcode in the ambient profile.

round
A logical scalar indicating whether to check for non-integer values in m and, if present, round them for ambient profile estimation (see ?ambientProfileEmpty) and the multinomial simulations.

niters
An integer scalar specifying the number of iterations to use for the Monte Carlo p-value calculations.

BPPARAM
A BiocParallelParam object indicating whether parallelization should be used.

cand.max.n
An integer scalar specifying the maximum number of barcodes that can be included in the cell candidate pool.

Value

A DataFrame with the same fields as that returned by emptyDrops.

Details about emptyDropsCellRanger arguments

The arguments in STARsolo are one-to-one correspond to the arguments in emptyDropsCellRanger:

n.expected.cells: This argument is used to define the number of expected cells in a sample. It is the same as nExpectedCells in STARsolo.

max.percentile: Together with n.expected.cells, max.percentile defines the maximum UMI count that an expected cell can have. This argument is same as maxPercentile in STARsolo.

max.min.ratio: Based on the maximum UMI count computed using n.expected.cells and max.percentile, max.min.ratio defines the threshold of the minimum UMI count that an expected cell can have according to their ratio. The barcodes whose UMI count is above this threshold are assigned as real cells by the simple filtering algorithm (nCellsSimple in STARsolo). It is same as maxMinRatio in STARsolo.

umi.min: This argument is one of the three thresholds used to define the lower bound of the cell candidate pool. umi.min specifies the minimum UMI count for inclusion of a barcode in the cell candidate pool. This argument is the same as umiMin in STARsolo.

umi.min.frac.median: This argument is one of the three thresholds used to define the lower bound of the cell candidate pool. This argument defines the minimum UMI as umi.min.frac.median times the median UMI count of the real cells assigned by the simple filtering algorithm. This argument is the same as umiMinFracMedian in STARsolo.

cand.max.n: This argument is one of the three thresholds used to define the lower bound of the cell candidate pool. This argument defines the minimum UMI as the UMI count of the barcodes whose UMI count ranks as the cand.max.n-th after the real cells defined by the simple filtering algorithm. This argument is the same as umiMinFracMedian in STARsolo.

ind.min: This argument specifies the lowest UMI count ranking of the barcodes in the ambient pool. This argument is same as indMin in STARsolo.

ind.max: This argument specifies the highest UMI count ranking of the barcodes in the ambient pool. This argument is same as indMin in STARsolo.

References

Kaminow et al. (2021). STARsolo: accurate, fast and versatile mapping/quantification of single-cell and single-nucleus RNA-seq data https://www.biorxiv.org/content/10.1101/2021.05.05.442755v1

See also

Author

Dongze He, Rob Patro

Examples

# Mocking up some data:
set.seed(0)
my.counts <- DropletUtils:::simCounts()

# Identify likely cell-containing droplets.
out <- emptyDropsCellRanger(my.counts)

#> Error in ambient.FUN(m, totals): no counts available to estimate the ambient profile

out

#> Error in eval(expr, envir, enclos): object 'out' not found


is.cell <- out$FDR <= 0.01

#> Error in eval(expr, envir, enclos): object 'out' not found

sum(is.cell, na.rm=TRUE)

#> Error in eval(expr, envir, enclos): object 'is.cell' not found


# Subsetting the matrix to the cell-containing droplets.
# (using 'which()' to handle NAs smoothly).
cell.counts <- my.counts[,which(is.cell),drop=FALSE]

#> Error in h(simpleError(msg, call)): error in evaluating the argument 'j' in selecting a method for function '[': error in evaluating the argument 'x' in selecting a method for function 'which': object 'is.cell' not found

dim(cell.counts)

#> Error in eval(expr, envir, enclos): object 'cell.counts' not found