Get 10x cell statistics

Compute some basic per-cell statistics from the 10x molecule information file.

get10xMolInfoStats(sample, barcode.length=NULL, use.library=NULL)

Arguments

sample
A string containing the path to the molecule information HDF5 file.

barcode.length
An integer scalar specifying the length of the cell barcode, see read10xMolInfo.

use.library
An integer vector specifying the library indices for which to extract molecules from sample. Alternatively, a character vector specifying the library type(s), e.g., "Gene expression".

Value

A DataFrame containing one row per cell library, with the fields:

cell:

Character, the cell barcode.

gem_group:

Integer, the GEM group.

num.umis:

Integer, the number of UMIs assigned to this cell barcode/GEM group combination.

num.reads:

Integer, the number of reads for this combination.

num.genes:

Integer, the number of detected genes.

Author

Aaron Lun

See also

Examples

# Mocking up some 10X HDF5-formatted data.
out <- DropletUtils:::simBasicMolInfo(tempfile())

get10xMolInfoStats(out)    

#> DataFrame with 256 rows and 5 columns
#>            cell gem_group  num.umis num.reads num.genes
#>     <character> <integer> <integer> <integer> <integer>
#> 1          AAAA         1        30       310        15
#> 2          AAAC         1        33       341        15
#> 3          AAAG         1        39       361        18
#> 4          AAAT         1        36       370        18
#> 5          AACA         1        50       493        17
#> ...         ...       ...       ...       ...       ...
#> 252        TTGT         1        36       314        17
#> 253        TTTA         1        38       388        18
#> 254        TTTC         1        42       395        16
#> 255        TTTG         1        47       446        19
#> 256        TTTT         1        38       379        18